refatorando classes

+++ src/main/java/temp/desafio/api/infrastructure/service/DadosMetereologicosService.java
@@ -6,10 +6,12 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import temp.desafio.api.core.dadosMetereologico.dto.DadosMetereologicos;
 import temp.desafio.api.core.dadosMetereologico.repositories.IDadosMetereologicos;
+import temp.desafio.api.core.enums.TipoTurno;
 import temp.desafio.api.infrastructure.mappers.DadosMetereologicosEntityMapper;
 import temp.desafio.api.infrastructure.persistence.entity.DadosMetereologicosEntity;
 import temp.desafio.api.infrastructure.persistence.repositories.DadosMetereologicosRepository;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -28,8 +30,8 @@ public class DadosMetereologicosService implements IDadosMetereologicos {
     @Override
     public DadosMetereologicos createDadosMetereologicos(DadosMetereologicos dadosMetereologicos){
         DadosMetereologicosEntity dadosMetereologicosEntity = entityMapper.toDadosMetereologicosEntity(dadosMetereologicos);
-        DadosMetereologicosEntity novoDadosMetereologicos = dadosMetereologicosRepository.save(dadosMetereologicosEntity);
-        return entityMapper.toDadosMetereologicos(novoDadosMetereologicos);
+        DadosMetereologicosEntity dadosMetereologicosEntitySaveEntity = dadosMetereologicosRepository.save(dadosMetereologicosEntity);
+        return entityMapper.toDadosMetereologicos(dadosMetereologicosEntitySaveEntity);
     }
 
     @Override
@@ -48,13 +50,40 @@ public class DadosMetereologicosService implements IDadosMetereologicos {
     }
 
     @Override
-    public void deleteDadosMeterologicos(DadosMetereologicos dadosMetereologicos) {
-        DadosMetereologicosEntity dadosMetereologicosEntity = entityMapper.toDadosMetereologicosEntity(dadosMetereologicos);
-        dadosMetereologicosRepository.delete(dadosMetereologicosEntity);
+    public void deleteDadosMeterologicos(
+            String cidade,
+            LocalDate data,
+            TipoTurno turno
+    ) {
+        Optional<DadosMetereologicosEntity> optionalDadosMetereologicosEntity = dadosMetereologicosRepository.findByCidadeAndDataAndTurno(cidade, data, turno);
+        if (optionalDadosMetereologicosEntity.isPresent()){
+            DadosMetereologicosEntity dadosMetereologicosEntity = optionalDadosMetereologicosEntity.get();
+            dadosMetereologicosRepository.delete(dadosMetereologicosEntity);
+        }
     }
 
     @Override
-    public DadosMetereologicos updateDadosMeterologicos(DadosMetereologicos dadosMetereologicos) {
-        return null;
+    public DadosMetereologicos updateDadosMeterologicos(
+            String cidade,
+            LocalDate data,
+            TipoTurno turno,
+            DadosMetereologicos dadosMetereologicos) {
+            Optional<DadosMetereologicosEntity> dadosMetereologicosEntity = dadosMetereologicosRepository.findByCidadeAndDataAndTurno(cidade, data, turno);
+        if (dadosMetereologicosEntity.isPresent()) {
+            DadosMetereologicosEntity existingDadosMetereologicosEntity = dadosMetereologicosEntity.get();
+            existingDadosMetereologicosEntity.setCidade(dadosMetereologicos.cidade());
+            existingDadosMetereologicosEntity.setData(dadosMetereologicos.data());
+            existingDadosMetereologicosEntity.setClima(dadosMetereologicos.clima());
+            existingDadosMetereologicosEntity.setTurno(dadosMetereologicos.turno());
+            existingDadosMetereologicosEntity.setPrecipitacao(dadosMetereologicos.precipitacao());
+            existingDadosMetereologicosEntity.setTurno(dadosMetereologicos.turno());
+            existingDadosMetereologicosEntity.setTempMax(dadosMetereologicos.tempMax());            existingDadosMetereologicosEntity.setTempMin(dadosMetereologicos.tempMin());
+            existingDadosMetereologicosEntity.setUmidade(dadosMetereologicos.umidade());
+            existingDadosMetereologicosEntity.setVelDoVento(dadosMetereologicos.velDoVento());
+            dadosMetereologicosRepository.save(existingDadosMetereologicosEntity);
+            return entityMapper.toDadosMetereologicos(existingDadosMetereologicosEntity);
+        } else {
+            throw new RuntimeException("DadosMetereologicos not found for cidade: " + cidade);
+        }
     }
 }
